WIP refactor(parser): simplified grammar rules for element attributes
Process the attributes in a more generic way: positional values
are retained until the block is initialized, in which case they
are process accordingly.
Also, no need to the `AdmonitionKind` and `BlockKind` types, whose
constant values are all `string` now and stored in the `style` key.

Also, for inline attributes, spaces after the opening squre bracket (`[`)
are not tolerated (this was already the case for block attributes). This
allows for a better consistency and simpler grammar rules.

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>